% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_connection}
\alias{check_connection}
\title{Check connection validity}
\usage{
check_connection(wrds)
}
\arguments{
\item{wrds}{Object to check.}
}
\value{
Invisibly returns \code{TRUE} if valid; aborts otherwise.
}
\description{
Internal function to validate a database connection object.
}
\keyword{internal}
