% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-company.R
\name{company_config}
\alias{company_config}
\title{Get company table configuration}
\usage{
company_config(region)
}
\arguments{
\item{region}{One of \code{"na"} or \code{"global"}.}
}
\value{
A list with \code{table} and \code{columns}.
}
\description{
Internal function returning table names and default columns for company
queries.
}
\keyword{internal}
