% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-compustat.R
\name{compustat_config}
\alias{compustat_config}
\title{Get Compustat configuration}
\usage{
compustat_config(frequency, region)
}
\arguments{
\item{frequency}{One of \code{"annual"} or \code{"quarterly"}.}

\item{region}{One of \code{"na"} or \code{"global"}.}
}
\value{
A list with \code{table}, \code{datafmt}, \code{popsrc}, and \code{columns}.
}
\description{
Internal function returning table names and filters for Compustat queries.
}
\keyword{internal}
