% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_libraries}
\alias{list_libraries}
\title{List available libraries}
\usage{
list_libraries(wrds)
}
\arguments{
\item{wrds}{A \code{DBIConnection} object returned by \code{\link[=wrds_connect]{wrds_connect()}}.}
}
\value{
A character vector of library names.
}
\description{
Returns a character vector of available schema names (libraries) on WRDS.
}
\examples{
\dontrun{
wrds <- wrds_connect()
list_libraries(wrds)
wrds_disconnect(wrds)
}
}
