% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds-package.R
\docType{package}
\name{wrds-package}
\alias{wrds}
\alias{wrds-package}
\title{wrds: Access 'Wharton Research Data Services' ('WRDS')}
\description{
Provides simple functions for accessing data from 'Wharton Research Data Services' ('WRDS'), a widely used financial database in academic research. Includes credential management via the system keyring, database tools, and functions for downloading generic tables, 'Compustat' fundamentals, and linking tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statzhero/wrds}
  \item Report bugs at \url{https://github.com/statzhero/wrds/issues}
}

}
\author{
\strong{Maintainer}: Ulrich Atz \email{ulrich.atz@unibocconi.it} (\href{https://orcid.org/0000-0002-1719-3780}{ORCID}) [copyright holder]

}
\keyword{internal}
