% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrds-connect.R
\name{wrds_disconnect}
\alias{wrds_disconnect}
\title{Disconnect from WRDS}
\usage{
wrds_disconnect(wrds)
}
\arguments{
\item{wrds}{A \code{DBIConnection} object returned by \code{\link[=wrds_connect]{wrds_connect()}}.}
}
\value{
Invisibly returns \code{TRUE} if disconnection was successful.
}
\description{
Closes a WRDS database connection.
}
\examples{
\dontrun{
wrds <- wrds_connect()
wrds_disconnect(wrds)
}
}
